fix(sort): Fix multi-sort with nils (#7432)
Fix the behavior of multi-sort. All the nil values are appended at the end of
the result irrespective of ascending or descending sort. This change also
fixes a bug due to not appending the nil values in the values list, 
corresponding to the UIDs appended in the UID list with nil predicates.