feat(GraphQL):This PR adds subscriptions to custom DQL.  (#7385)
Fixes GRAPHQL-993
As of now, we only allow subscriptions on GraphQL queries. As we can use DQL through @custom in graphql, many users requested to use subscriptions on custom DQL.

We have the @withSubscription directive, that we use on GraphQL types to generate subscriptions on queries for that type.
We are now extending this directive to DQL queries in custom queries. Users can specify @withSubscription directive on individual DQL queries in type Query and those queries will be added in type subscription.

For example, we have two custom DQL queries below queryTweetsSortedByAuthorFollowers and queryUserTweetCounts . queryUserTweetCounts have @withSubscription directive, and it will be added to subscription type and users will be able to subscribe to this query.

type Query {
  queryTweetsSortedByAuthorFollowers(search: String!): [Tweets] @custom(dql: """
	query q($search: string) {
		var(func: type(Tweets)) @filter(anyoftext(Tweets.text, $search)) {
			Tweets.author {
				followers as User.followers
			}
			authorFollowerCount as sum(val(followers))
		}
		queryTweetsSortedByAuthorFollowers(func: uid(authorFollowerCount), orderdesc: val(authorFollowerCount)) {
			id: uid
			text: Tweets.text
			author: Tweets.author {
			    screen_name: User.screen_name
			    followers: User.followers
			}
			timestamp: Tweets.timestamp
		}
	}
	""")

  queryUserTweetCounts: [UserTweetCount]  `@withSubscription` @custom(dql: """
	query {
		queryUserTweetCounts(func: type(User)) {
			screen_name: User.screen_name
			tweetCount: count(User.tweets)
		}
	}
	""")
}