Feat(GraphQL): This PR allow multiple @id fields in a type. (#7235)
Fixes GRAPHQL-924
Currently, we only allow single xid or external id fields in a type with the @id directive.
Now, we are extending this to have multiple external @id fields in a type. Every such field will be unique over all the nodes of that type. As of now, fields with @id directive will be immutable. get query will work for all the fields with the @id type.
RFC:https://discuss.dgraph.io/t/wip-rfc-allow-multiple-unique-fields-in-graphql-schema/12008
Co-authored-by: Rajas Vanjape <rajas@dgraph.io>