fix(DQL): optimize query for has function with offset. (#7727)
Fixes DGRAPH-574.

This PR reduces the latency of the `has` function slightly. For example the given query:-
```
{
    me(func: has(actor.film), first: 10, offset: 300000){
       name@en
    }
}
```
when run on the `21million` dataset takes around `80` milliseconds to execute as compared to earlier around `93` milliseconds.

Machine Specs:-
```
Model Name:	MacBook Pro
  Model Identifier:	MacBookPro16,1
  Processor Name:	6-Core Intel Core i7
  Processor Speed:	2.6 GHz
  Number of Processors:	1
  Total Number of Cores:	6
  L2 Cache (per Core):	256 KB
  L3 Cache:	12 MB
  Memory:	16 GB
```

Note:-
This does not optimize for the `has` function on predicates with the `@lang` tag.