fix(live): make live loader progress on a cluster with very high maxUid  (#7743)
This PR fixes the following case:
If we have a cluster that cannot lease out new UIDs because it has already leased upto its
max limit. Now, we try to live load the data with the given UIDs and the AssignIds complains
that the limit has reached. Hence, update the xidmap's maxSeenUid and make progress.