opt(snapshot): optimize snapshot by using sinceTs (#7826)
We were not using stream.SinceTs in snapshot transfer. Instead, we were using ChooseKey to skip the keys with a lesser version. This can be very slow and memory heavy, because we will have to iterate over the entire DB and skip the keys. Using SinceTs, we will be skipping the entire table.