feat(schema): do schema versioning and make backup non-blocking for indexing (#7852)
Issue:
We used to write schema/types at timestamp=1. That single bit violates the Snapshot Isolation. Hence, we couldn't run schema updates and other operations like backup concurrently. This is an issue in a multi-tenant cluster because if the backup is running we couldn't update the schema.
Fix:
This PR writes the schema/types at the startTs of the transaction and enables the schema updates to be made even if the backup is running.