Feat(Backup): Add native google cloud storage backup support (#7829)
Currently, we use MinIO to write backup files to CGP storage. To do this we need to start a minio server (with GCP credentials) and trigger a backup with the destination as the minio server. This requires the user to run the minio server every time (or always) when a backup needs to be triggered. Using native SDK will help get rid of the extra service.

Google cloud storage (GCS) requires the user to create Service account key and credentials file. When starting an alpha we can set env variable:

GOOGLE_APPLICATION_CREDENTIALS="/path/to/credentials.json" dgraph alpha

to allow access to GCS. To trigger a backup the user can use the admin endpoint to initiate the backup. The destination should be set as:

"gs://<bucket_name>/path/to/backup"