opt(query): Use sroar in pb.List (#7864)
Use roaring bitmaps in pb.List by default. We fall back to using sorted_uids 
when we need the list to be ordered. The APIs defined in the codec acts as
an abstraction layer that decides if sorted_uids or bitmap should be used. 
sorted_uids are used if it is present, otherwise, the bitmap is used.

Co-authored-by: Manish R Jain <manish@dgraph.io>