fix(snapshot): update last snapshot time across members (#7968)
We see frequent snapshots in an HA cluster because of the leadership re-election because of leader crash/shutdown. The local variable that we use for snapshot frequency is not synchronized across alphas. So each alpha thinks differently about the time the last snapshot was taken.

(cherry picked from commit 43055bf0d3b483da93e604e64eb6ca02566d5921)