fix(graphql): fix type assertion failure in graphql if resolver is not defined (#8003)
We were not checking the result of type assertion in lambda mutation's response. This causes the GraphQL server to trap a panic. Instead, it should have returned an empty response for the unresolved field.
Also, it stringifies the logs from the lambda server.
Some minor deduplication in the posting list code.