fix(lambda): shutdown node processes when alpha gets killed (#8027)
We were already handling the graceful shutdown of node processes when alpha shuts down. We were passing the closer.Ctx() so that node process shuts down when alpha stops. But that is not sufficient when alpha panics.
In case alpha panics, the alpha gets stuck because the node processes. This happens when sentry is also enabled. Sentry waits for all processes (child as well as grandchild to complete).