feat(magicNumber): Introduce magic number (#8032)
Magic number is a unique identifier for the data format of dgraph. In 21.09 we
have changed the data format of posting lists by bringing in sroar. Running
Dgraph with sroar change on older p directory can cause data corruption.
This magic number prevents DB to start up if the version is not compatible.